Java MinAvgTwoSlice && MaxDoubleSliceSum. C++ macCounters

+++ 5 PrefixSums/MinAvgTwoSlice/minAvgTwoSlice.java
@@ -4,12 +4,13 @@ class Solution {
     double avg = 0;
     double minAvg = (A[0] + A[1]) / 2;
     for (int i = 2; i < A.length; i += 1) {
-      avg = (double) (A[i - 2] + A[i - 1] + A[i]) / 3;
+      avg = (A[i - 2] + A[i - 1] + A[i]) / 3.0;
       if (minAvg > avg) {
         minAvg = avg;
         output = i - 2;
       }
-      avg = (double) (A[i - 1] + A[i]) / 2;
+      avg = (A[i - 1] + A[i]) / 2.0;
+      //avg = (double) (A[i - 1] + A[i]) / 2;
       if (minAvg > avg) {
         minAvg = avg;
         output = i - 1;
